Expose a FileSystem function to write the configuration
